pyproject.toml
[project]
name = "distill-rss"
version = "0.1.0"
description = "AI-powered RSS reader with Gemini analysis and MCP tool integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "google-genai>=0.1.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.3.3",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

